added get_nft_assets_owners
