--- scripts/actions/get-metanode-subscription-account.js
@@ -1,79 +1,0 @@
-'use strict';
-// libraries
-
-// modules
-
-// constants
-const ACTION = 'get_metanode_subscription_account';
-
-// variables
-/* eslint-disable no-unused-vars */
-let config;
-let loggingUtil;
-/* eslint-enable no-unused-vars */
-
-// functions
-const init = (_config, _loggingUtil) => {
-  /* istanbul ignore if */
-  if (_config === undefined) {
-    throw new Error('config is required.');
-  }
-  /* istanbul ignore if */
-  if (_loggingUtil === undefined) {
-    throw new Error('loggingUtil is required.');
-  }
-  config = _config;
-  loggingUtil = _loggingUtil;
-};
-
-const deactivate = () => {
-  /* eslint-disable no-unused-vars */
-  config = undefined;
-  loggingUtil = undefined;
-  /* eslint-enable no-unused-vars */
-};
-
-/**
- * gets the bananode url
- * @memberof Main
- * @param {Object} context the context, used to get cached data.
- * - from filesystem in nodejs,
- * - from localstorage in a browser,
- * - from a test harness in the unit tests.
- * @param {Object} req the http request.
- * @param {Object} res the http response.
- * @return {undefined}
- */
-const getBananodeApiUrl = async (context, req, res) => {
-  const fetch = context.fetch;
-
-  /* istanbul ignore if */
-  if (fetch === undefined) {
-    throw Error('context.fetch is required');
-  }
-
-  /* istanbul ignore if */
-  if (req === undefined) {
-    throw Error('req is required');
-  }
-
-  /* istanbul ignore if */
-  if (req.body === undefined) {
-    throw Error('req.body is required');
-  }
-
-  const resp = {};
-  resp.success = true;
-  resp.metanode_subscription_account = config.metanodeSubscriptionAccount;
-
-  res.send(resp);
-};
-
-const addAction = (actions) => {
-  actions[ACTION] = getBananodeApiUrl;
-};
-
-exports.init = init;
-exports.deactivate = deactivate;
-exports.ACTION = ACTION;
-exports.addAction = addAction;
