+++ scripts/config.json
@@ -1,5 +1,4 @@
 {
-  "metanodeSubscriptionAccount":"ban_1metanodesubscription1111111111111111111111111111111hx48xsub",
   "bananodeApiUrl": "https://kaliumapi.appditto.com/api",
   "webPort": ""
 }
