+++ static-html/index.html
@@ -16,7 +16,9 @@
     </div>
     <div id="mintNftWrapper" class="container column nowrap">
     </div>
-    <div id="metanodeSubscriptionWrapper" class="container column nowrap">
+    <div id="ownerCheckWrapper" class="container column nowrap">
+    </div>
+    <div id="ownerChangeWrapper" class="container column nowrap">
     </div>
   </div>
   <script type="module" src="index.js"></script>
